refactor: Update input values in demande-list and edit-demande components
The input values in the demande-list and edit-demande components have been updated to correctly set the depart and destination properties of the newDemande and demande objects respectively. This change ensures that the correct values are assigned when a suggestion item is clicked, improving the accuracy of the data and user experience.